proper grayscale upon conversion
by removing the over/underflow bins (which are all 0) from the output
TASImage, the grayscale now works properly, with the minimum actually
represented by black and the maximum by white.

Before this change, the black bins where mapped to a light gray as they
contained some baseline counts and the minimum in the image was 0.