can export datasets from hdf5 files
dpc_radiography now saves the drift corrected images in the postprocessing
group of the hdf5 file.
A new method (save_images) takes care of that.

export_dataset takes an hdf5 file and the path of
one or more datasets and saves the
results as images.

this script is added to the setup.py